app/requirements.txt
anyio==3.7.0
async-generator==1.10
attrs==23.1.0
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
colorama==0.4.6
fastapi==0.95.2
fastapi-login==1.9.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql==0.0.3
mysql-connector-python==8.0.33
mysqlclient==2.1.1
outcome==1.2.0
passlib==1.7.4
protobuf==3.20.3
pycparser==2.21
pydantic==1.10.8
PyJWT==2.7.0
python-dotenv==1.0.0
python-multipart==0.0.6
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.15
starlette==0.27.0
trio==0.21.0
typing_extensions==4.6.3
uvicorn==0.22.0
Werkzeug==2.3.4
